requirements.txt
# Core dependencies
fastapi
uvicorn
streamlit
python-dotenv
requests
pydantic

# AI and News API dependencies
agno
newsapi-python


asyncio-throttle

# Additional utilities
python-multipart
aiofiles    